requirements.txt
python-chess
stockfish
numpy
transformers
accelerate
torch
datasets
